feat: add upload files mutations
