+++ src/item/api.ts
@@ -9,25 +9,20 @@ import {
 } from '@graasp/sdk';
 
 import { verifyAuthentication } from '../api/axios.js';
-import { DEFAULT_THUMBNAIL_SIZE } from '../config/constants.js';
 import { PartialQueryConfigForApi } from '../types.js';
 import {
   GET_OWN_ITEMS_ROUTE,
   GET_RECYCLED_ITEMS_DATA_ROUTE,
   SHARED_ITEM_WITH_ROUTE,
   buildCopyItemsRoute,
-  buildDeleteItemThumbnailRoute,
   buildDeleteItemsRoute,
   buildDownloadFilesRoute,
-  buildDownloadItemThumbnailRoute,
   buildEditItemRoute,
   buildGetChildrenRoute,
   buildGetItemParents,
   buildGetItemRoute,
   buildGetItemsRoute,
   buildMoveItemsRoute,
-  buildPostItemRoute,
-  buildPostItemWithThumbnailRoute,
   buildRecycleItemsRoute,
   buildRestoreItemsRoute,
 } from './routes.js';
@@ -64,89 +59,11 @@ export type PostItemPayloadType = Partial<DiscriminatedItem> &
   Partial<{
     parentId: UUID;
     geolocation: Pick<ItemGeolocation, 'lat' | 'lng'>;
+    previousItemId: UUID;
   }>;
 export type PostItemWithThumbnailPayloadType = PostItemPayloadType & {
   thumbnail: Blob;
 };
-
-// payload = {name, type, description, extra, geolocation}
-// querystring = {parentId}
-export const postItem = async (
-  {
-    name,
-    displayName,
-    type,
-    description,
-    extra,
-    parentId,
-    geolocation,
-    settings,
-  }: PostItemPayloadType,
-  { API_HOST, axios }: PartialQueryConfigForApi,
-): Promise<DiscriminatedItem> =>
-  verifyAuthentication(() =>
-    axios
-      .post<DiscriminatedItem>(`${API_HOST}/${buildPostItemRoute(parentId)}`, {
-        name: name.trim(),
-        displayName: displayName?.trim() ?? name.trim(),
-        type,
-        description,
-        extra,
-        geolocation,
-        settings,
-      })
-      .then(({ data }) => data),
-  );
-
-export const postItemWithThumbnail = async (
-  {
-    name,
-    displayName,
-    type,
-    description,
-    extra,
-    parentId,
-    geolocation,
-    settings,
-    thumbnail,
-  }: PostItemWithThumbnailPayloadType,
-  { API_HOST, axios }: PartialQueryConfigForApi,
-): Promise<DiscriminatedItem> =>
-  verifyAuthentication(() => {
-    const itemPayload = new FormData();
-    // name and type are required
-    itemPayload.append('name', name);
-    itemPayload.append('type', type);
-    if (displayName) {
-      itemPayload.append('displayName', displayName);
-    }
-    if (description) {
-      itemPayload.append('description', description);
-    }
-    if (geolocation) {
-      itemPayload.append('geolocation', JSON.stringify(geolocation));
-    }
-    if (settings) {
-      itemPayload.append('settings', JSON.stringify(settings));
-    }
-    if (extra) {
-      itemPayload.append('extra', JSON.stringify(extra));
-    }
-    /* WARNING: this file field needs to be the last one,
-     * otherwise the normal fields can not be read
-     * https://github.com/fastify/fastify-multipart?tab=readme-ov-file#usage
-     */
-    itemPayload.append('file', thumbnail);
-    return axios
-      .post<DiscriminatedItem>(
-        `${API_HOST}/${buildPostItemWithThumbnailRoute(parentId)}`,
-        itemPayload,
-        {
-          headers: { 'Content-Type': 'multipart/form-data' },
-        },
-      )
-      .then(({ data }) => data);
-  });
 
 export const deleteItems = async (
   ids: UUID[],
@@ -309,42 +226,3 @@ export const restoreItems = async (
       .post<void>(`${API_HOST}/${buildRestoreItemsRoute(itemIds)}`)
       .then(({ data }) => data),
   );
-
-export const downloadItemThumbnail = async (
-  { id, size = DEFAULT_THUMBNAIL_SIZE }: { id: UUID; size?: string },
-  { API_HOST, axios }: PartialQueryConfigForApi,
-) =>
-  axios
-    .get<Blob>(
-      `${API_HOST}/${buildDownloadItemThumbnailRoute({
-        id,
-        size,
-        replyUrl: false,
-      })}`,
-      {
-        responseType: 'blob',
-      },
-    )
-    .then(({ data }) => data);
-
-export const downloadItemThumbnailUrl = async (
-  { id, size = DEFAULT_THUMBNAIL_SIZE }: { id: UUID; size?: string },
-  { API_HOST, axios }: PartialQueryConfigForApi,
-) =>
-  axios
-    .get<string>(
-      `${API_HOST}/${buildDownloadItemThumbnailRoute({
-        id,
-        size,
-        replyUrl: true,
-      })}`,
-    )
-    .then(({ data }) => data);
-
-export const deleteItemThumbnail = async (
-  id: UUID,
-  { API_HOST, axios }: PartialQueryConfigForApi,
-) =>
-  axios
-    .delete<void>(`${API_HOST}/${buildDeleteItemThumbnailRoute(id)}`)
-    .then(({ data }) => data);
