+++ src/item/hooks.test.ts
@@ -5,7 +5,6 @@ import {
   ItemType,
   LocalFileItemFactory,
   MAX_TARGETS_FOR_READ_REQUEST,
-  ThumbnailSize,
 } from '@graasp/sdk';
 
 import { StatusCodes } from 'http-status-codes';
@@ -14,8 +13,6 @@ import { afterEach, describe, expect, it } from 'vitest';
 
 import {
   FILE_RESPONSE,
-  THUMBNAIL_BLOB_RESPONSE,
-  THUMBNAIL_URL_RESPONSE,
   UNAUTHORIZED_RESPONSE,
   buildResultOfData,
   generateFolders,
@@ -30,7 +27,6 @@ import { OWN_ITEMS_KEY, itemKeys } from '../keys.js';
 import {
   GET_OWN_ITEMS_ROUTE,
   buildDownloadFilesRoute,
-  buildDownloadItemThumbnailRoute,
   buildGetChildrenRoute,
   buildGetItemParents,
   buildGetItemRoute,
@@ -549,256 +545,6 @@ describe('useFileContent', () => {
 
     expect(isError).toBeTruthy();
     expect(data).toBeFalsy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-});
-
-describe('useItemThumbnail', () => {
-  afterEach(() => {
-    nock.cleanAll();
-    queryClient.clear();
-  });
-
-  const item = FolderItemFactory();
-  const replyUrl = false;
-  const key = itemKeys.single(item.id).thumbnail({ replyUrl });
-  const response = THUMBNAIL_BLOB_RESPONSE;
-  const route = `/${buildDownloadItemThumbnailRoute({
-    id: item.id,
-    replyUrl,
-  })}`;
-  const hook = () => hooks.useItemThumbnail({ id: item.id });
-
-  it(`Receive default thumbnail`, async () => {
-    const endpoints = [
-      {
-        route,
-        response,
-        headers: {
-          'Content-Type': 'image/jpeg',
-        },
-      },
-    ];
-    const { data } = await mockHook({ endpoints, hook, wrapper });
-
-    expect(data).toBeTruthy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeTruthy();
-  });
-
-  it(`Receive large thumbnail`, async () => {
-    const size = ThumbnailSize.Large;
-    const routeLarge = `/${buildDownloadItemThumbnailRoute({
-      id: item.id,
-      size,
-      replyUrl,
-    })}`;
-    const hookLarge = () => hooks.useItemThumbnail({ id: item.id, size });
-    const keyLarge = itemKeys.single(item.id).thumbnail({ size });
-
-    const endpoints = [
-      {
-        route: routeLarge,
-        response,
-        headers: {
-          'Content-Type': 'image/jpeg',
-        },
-      },
-    ];
-    const { data } = await mockHook({
-      endpoints,
-      hook: hookLarge,
-      wrapper,
-    });
-
-    expect(data).toBeTruthy();
-    // verify cache keys
-    expect(queryClient.getQueryData(keyLarge)).toBeTruthy();
-  });
-
-  it(`Undefined id does not fetch`, async () => {
-    const endpoints = [
-      {
-        route,
-        response,
-      },
-    ];
-    const { data, isFetched } = await mockHook({
-      endpoints,
-      hook: () => hooks.useItemThumbnail({ id: undefined }),
-      wrapper,
-      enabled: false,
-    });
-
-    expect(data).toBeFalsy();
-    expect(isFetched).toBeFalsy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-
-  it(`Does not fetch if item has no thumbnail`, async () => {
-    const itemWithoutThumbnail = {
-      ...item,
-      settings: { hasThumbnail: false },
-    };
-    queryClient.setQueryData(
-      itemKeys.single(itemWithoutThumbnail.id).content,
-      itemWithoutThumbnail,
-    );
-    const endpoints = [
-      {
-        route,
-        response,
-      },
-    ];
-    const { data, isFetched } = await mockHook({
-      endpoints,
-      hook: () => hooks.useItemThumbnail({ id: itemWithoutThumbnail.id }),
-      wrapper,
-      enabled: false,
-    });
-
-    expect(data).toBeFalsy();
-    expect(isFetched).toBeFalsy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-
-  it(`Unauthorized`, async () => {
-    const endpoints = [
-      {
-        route,
-        response: UNAUTHORIZED_RESPONSE,
-        statusCode: StatusCodes.UNAUTHORIZED,
-      },
-    ];
-    const { data, isError } = await mockHook({ endpoints, hook, wrapper });
-
-    expect(data).toBeFalsy();
-    expect(isError).toBeTruthy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-});
-
-describe('useItemThumbnailUrl', () => {
-  afterEach(() => {
-    nock.cleanAll();
-    queryClient.clear();
-  });
-
-  const item = FolderItemFactory();
-  const replyUrl = true;
-  const key = itemKeys.single(item.id).thumbnail({ replyUrl });
-  const response = THUMBNAIL_URL_RESPONSE;
-  const route = `/${buildDownloadItemThumbnailRoute({
-    id: item.id,
-    replyUrl,
-  })}`;
-  const hook = () => hooks.useItemThumbnailUrl({ id: item.id });
-
-  it(`Receive default thumbnail`, async () => {
-    const endpoints = [
-      {
-        route,
-        response,
-      },
-    ];
-    const { data } = await mockHook({ endpoints, hook, wrapper });
-
-    expect(data).toBeTruthy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeTruthy();
-  });
-
-  it(`Receive large thumbnail`, async () => {
-    const size = ThumbnailSize.Large;
-    const routeLarge = `/${buildDownloadItemThumbnailRoute({
-      id: item.id,
-      size,
-      replyUrl,
-    })}`;
-    const hookLarge = () => hooks.useItemThumbnailUrl({ id: item.id, size });
-    const keyLarge = itemKeys.single(item.id).thumbnail({ size, replyUrl });
-
-    const endpoints = [
-      {
-        route: routeLarge,
-        response,
-      },
-    ];
-    const { data } = await mockHook({
-      endpoints,
-      hook: hookLarge,
-      wrapper,
-    });
-
-    expect(data).toBeTruthy();
-    // verify cache keys
-    expect(queryClient.getQueryData(keyLarge)).toBeTruthy();
-  });
-
-  it(`Undefined id does not fetch`, async () => {
-    const endpoints = [
-      {
-        route,
-        response,
-      },
-    ];
-    const { data, isFetched } = await mockHook({
-      endpoints,
-      hook: () => hooks.useItemThumbnail({ id: undefined }),
-      wrapper,
-      enabled: false,
-    });
-
-    expect(data).toBeFalsy();
-    expect(isFetched).toBeFalsy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-
-  it(`Does not fetch if item has no thumbnail`, async () => {
-    const itemWithoutThumbnail = {
-      ...item,
-      settings: { hasThumbnail: false },
-    };
-    queryClient.setQueryData(
-      itemKeys.single(itemWithoutThumbnail.id).content,
-      itemWithoutThumbnail,
-    );
-    const endpoints = [
-      {
-        route,
-        response,
-      },
-    ];
-    const { data, isFetched } = await mockHook({
-      endpoints,
-      hook: () => hooks.useItemThumbnail({ id: itemWithoutThumbnail.id }),
-      wrapper,
-      enabled: false,
-    });
-
-    expect(data).toBeFalsy();
-    expect(isFetched).toBeFalsy();
-    // verify cache keys
-    expect(queryClient.getQueryData(key)).toBeFalsy();
-  });
-
-  it(`Unauthorized`, async () => {
-    const endpoints = [
-      {
-        route,
-        response: UNAUTHORIZED_RESPONSE,
-        statusCode: StatusCodes.UNAUTHORIZED,
-      },
-    ];
-    const { data, isError } = await mockHook({ endpoints, hook, wrapper });
-
-    expect(data).toBeFalsy();
-    expect(isError).toBeTruthy();
     // verify cache keys
     expect(queryClient.getQueryData(key)).toBeFalsy();
   });
